Extract modal initialization to modal-init.js
Move modal registration logic from app.js to a dedicated module.
The initModals function now accepts (modals, terminal) for better
composability.

Reduces app.js by ~31 lines (854 → 823).
